refactor: various changes feat: re-render progress on calorie and weight change

+++ frontend/src/components/Footer/index.jsx
@@ -1,27 +1,43 @@
-import { Typography } from "@material-tailwind/react";
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { Typography } from '@material-tailwind/react';
 import icon from '../../assets/icon-flat.png';
 
 const YEAR = new Date().getFullYear();
 
-const LINKS = [
+const BASE_LINKS = [
     {
-        title: "Om os",
-        href: "#",
+        title: 'Om os',
+        href: '#',
     },
     {
-        title: "Kontakt os",
-        href: "#",
-    },
-    (localStorage.getItem('token') !== null ? {
-        title: "Log ud",
-        href: "/",
-        onClick: () => {
-            localStorage.clear();
-        }
-    } : {})
+        title: 'Kontakt os',
+        href: '#',
+    }
 ];
 
 export default function Footer() {
+    const [links, setLinks] = useState(BASE_LINKS);
+    const navigate = useNavigate();
+
+    useEffect(() => {
+        const isLoggedIn = localStorage.getItem('token') !== null;
+
+        const updatedLinks = [...BASE_LINKS];
+        if (isLoggedIn) {
+            updatedLinks.push({
+                title: 'Log ud',
+                href: '/',
+                onClick: () => {
+                    localStorage.clear();
+                    navigate('/', { replace: true })
+                }
+            });
+        }
+
+        setLinks(updatedLinks);
+    }, []);
+
     return (
         <footer className="w-full flex justify-center items-center p-4">
             <div>
@@ -30,16 +46,21 @@ export default function Footer() {
             <div className="w-full">
                 <div className="flex w-full flex-row flex-wrap items-center justify-end gap-x-12 gap-y-3 text-center">
                     <ul className="flex flex-wrap items-center gap-x-6 gap-y-2">
-                        {LINKS.map(({ title, href, onClick }, key) => (
+                        {links.map(({ title, href, onClick }, key) => (
                             <li key={key}>
-                                <Typography as="a" href={href || ''} onClick={onClick || (() => {})} className="text-foreground hover:text-primary">
+                                <Typography
+                                    as="a"
+                                    href={href}
+                                    onClick={onClick}
+                                    className="text-foreground hover:text-primary"
+                                >
                                     {title}
                                 </Typography>
                             </li>
                         ))}
                     </ul>
                 </div>
-                <hr className="my-4 border-surface"/>
+                <hr className="my-4 border-surface" />
                 <Typography className="text-center text-foreground">
                     &copy; {YEAR} FridgePilot co. FridgePilot og FridgePilot-logoet er blandt vores registrerede og uregistrerede varemærker i Danmark og andre lande.
                 </Typography>
